requirements.txt
google-adk
pyiceberg~=0.10.0
psycopg2-binary
protobuf~=6.33.1
python-dotenv~=1.2.1
SQLAlchemy~=2.0.44
trino